move natural gas pricing to feedstocks and convert config for compliance
